Refactor back navigation for multiplatform
